Move out add_node, add_edge, and related methods to RdfGraphMixin
- Moves add_node, add_edge, add_node_attribute, add_edge_attribute
methods out of RdfTransformer and into RdfGraphMixin
- Removes abstract class declaration from RdfTransformer
- Merge HgncRdfTransformer functionality directly with RdfTransformer
- Add a new kgx_utils.py for generic utility methods
- Tidy up code
- Document all classes and methods
