+++ kgx/__init__.py
@@ -4,8 +4,8 @@ __version__ = '0.0.1'
 
 from .pandas_transformer import PandasTransformer
 from .nx_transformer import GraphMLTransformer
+from .rdf_transformer import RdfTransformer, ObanRdfTransformer, RdfOwlTransformer
 from .sparql_transformer import SparqlTransformer, RedSparqlTransformer
-from .rdf_transformer import ObanRdfTransformer, RdfOwlTransformer, HgncRdfTransformer
 from .json_transformer import JsonTransformer
 from .neo_transformer import NeoTransformer
 from .logicterm_transformer import LogicTermTransformer
